Create RLEG_PGG_timeseries.f90
Here, we provide the source codes in Fortran to enable others to fully benefit from this research and replicate our findings. Specifically, the file named RLEG_PGG_timeseries.f90 contains the code that computes the time series of the global cooperation level within the population, given a specific game parameter of r = 0.72. The resulting data is then visualized in Fig.4(c).